+++ firebase-inappmessaging-display/src/main/java/com/google/firebase/inappmessaging/display/ktx/InAppMessagingDisplay.kt
@@ -27,7 +27,7 @@ import com.google.firebase.ktx.Firebase
  * respective main modules, and the Kotlin extension (KTX) APIs in
  * `com.google.firebase.firebase-inappmessaging-display-ktx` are now deprecated. As early as April
  * 2024, we'll no longer release KTX modules. For details, see the
- * [FAQ about this initiative.](https://firebase.google.com/docs/android/ktx-apis-to-main-modules)
+ * [FAQ about this initiative.](https://firebase.google.com/docs/android/kotlin-migration)
  */
 @Deprecated(
   "Use `com.google.firebase.Firebase.inAppMessagingDisplay` from the main module instead.",
@@ -49,7 +49,7 @@ val Firebase.inAppMessagingDisplay: FirebaseInAppMessagingDisplay
  * respective main modules, and the Kotlin extension (KTX) APIs in
  * `com.google.firebase.firebase-inappmessaging-display-ktx` are now deprecated. As early as April
  * 2024, we'll no longer release KTX modules. For details, see the
- * [FAQ about this initiative.](https://firebase.google.com/docs/android/ktx-apis-to-main-modules)
+ * [FAQ about this initiative.](https://firebase.google.com/docs/android/kotlin-migration)
  */
 @Deprecated(
   "Use `com.google.firebase.inappmessaging.display.FirebaseInAppMessagingDisplayKtxRegistrar` from the main module instead.",