+++ firebase-installations/src/main/java/com/google/firebase/installations/ktx/Installations.kt
@@ -26,7 +26,7 @@ import com.google.firebase.ktx.Firebase
  * respective main modules, and the Kotlin extension (KTX) APIs in
  * `com.google.firebase.firebase-storage-ktx` are now deprecated. As early as April 2024, we'll no
  * longer release KTX modules. For details, see the
- * [FAQ about this initiative.](https://firebase.google.com/docs/android/ktx-apis-to-main-modules)
+ * [FAQ about this initiative.](https://firebase.google.com/docs/android/kotlin-migration)
  */
 @Deprecated(
   "Use `com.google.firebase.Firebase.installations` from the main module instead.",
@@ -44,7 +44,7 @@ val Firebase.installations: FirebaseInstallations
  * respective main modules, and the Kotlin extension (KTX) APIs in
  * `com.google.firebase.firebase-storage-ktx` are now deprecated. As early as April 2024, we'll no
  * longer release KTX modules. For details, see the
- * [FAQ about this initiative.](https://firebase.google.com/docs/android/ktx-apis-to-main-modules)
+ * [FAQ about this initiative.](https://firebase.google.com/docs/android/kotlin-migration)
  */
 @Deprecated(
   "Use `com.google.firebase.Firebase.installations(app)` from the main module instead.",