+++ firebase-storage/src/main/java/com/google/firebase/storage/network/NetworkRequest.java
@@ -15,8 +15,6 @@
 package com.google.firebase.storage.network;
 
 import android.content.Context;
-import android.content.pm.PackageInfo;
-import android.content.pm.PackageManager;
 import android.net.ConnectivityManager;
 import android.net.Uri;
 import android.text.TextUtils;
@@ -28,6 +26,7 @@ import com.google.android.gms.common.internal.Preconditions;
 import com.google.android.gms.tasks.TaskCompletionSource;
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.emulators.EmulatedServiceSettings;
+import com.google.firebase.storage.BuildConfig;
 import com.google.firebase.storage.StorageException;
 import com.google.firebase.storage.internal.StorageReferenceUri;
 import com.google.firebase.storage.network.connection.HttpURLConnectionFactory;
@@ -50,23 +49,18 @@ import org.json.JSONObject;
 /** Encapsulates a single network request and response */
 @SuppressWarnings("unused")
 public abstract class NetworkRequest {
-  private static final String TAG = "NetworkRequest";
-
-  private static final String X_FIREBASE_GMPID = "x-firebase-gmpid";
-  private static final String X_FIREBASE_APPCHECK = "x-firebase-appcheck";
-
   public static final Uri PROD_BASE_URL = Uri.parse("https://firebasestorage.googleapis.com/v0");
-
   /* Do not change these values without changing corresponding logic on the SDK side*/
   public static final int INITIALIZATION_EXCEPTION = -1;
   public static final int NETWORK_UNAVAILABLE = -2;
-
   /*package*/ static final String GET = "GET";
   /*package*/ static final String DELETE = "DELETE";
   /*package*/ static final String POST = "POST";
   /*package*/ static final String PATCH = "PATCH";
   /*package*/ static final String PUT = "PUT";
-
+  private static final String TAG = "NetworkRequest";
+  private static final String X_FIREBASE_GMPID = "x-firebase-gmpid";
+  private static final String X_FIREBASE_APPCHECK = "x-firebase-appcheck";
   private static final int MAXIMUM_TOKEN_WAIT_TIME_MS = 30000;
   private static final String CONTENT_TYPE = "Content-Type";
   private static final String APPLICATION_JSON = "application/json";
@@ -80,7 +74,6 @@ public abstract class NetworkRequest {
 
   private StorageReferenceUri storageReferenceUri;
 
-  private static String gmsCoreVersion;
   private Context context;
   private Map<String, List<String>> resultHeaders;
   private int resultCode;
@@ -321,23 +314,6 @@ public abstract class NetworkRequest {
     return conn;
   }
 
-  @NonNull
-  private static String getGmsCoreVersion(Context context) {
-    if (gmsCoreVersion == null) {
-      PackageManager packageManager = context.getPackageManager();
-      try {
-        PackageInfo info = packageManager.getPackageInfo("com.google.android.gms", 0);
-        gmsCoreVersion = info.versionName;
-      } catch (PackageManager.NameNotFoundException e) {
-        Log.e(TAG, "Unable to find gmscore in package manager", e);
-      }
-      if (gmsCoreVersion == null) {
-        gmsCoreVersion = "[No Gmscore]";
-      }
-    }
-    return gmsCoreVersion;
-  }
-
   @SuppressWarnings("TryFinallyCanBeTryWithResources")
   private void constructMessage(
       @NonNull HttpURLConnection conn, @Nullable String authToken, @Nullable String appCheckToken)
@@ -355,14 +331,8 @@ public abstract class NetworkRequest {
     } else {
       Log.w(TAG, "No App Check token for request.");
     }
-
-    StringBuilder userAgent = new StringBuilder("Android/");
-    String gmsCore = getGmsCoreVersion(context);
-    if (!TextUtils.isEmpty(gmsCore)) {
-      userAgent.append(gmsCore);
-    }
-    conn.setRequestProperty("X-Firebase-Storage-Version", userAgent.toString());
-
+    String userAgent = "Android/" + BuildConfig.VERSION_NAME;
+    conn.setRequestProperty("X-Firebase-Storage-Version", userAgent);
     Map<String, String> requestProperties = requestHeaders;
     for (Map.Entry<String, String> entry : requestProperties.entrySet()) {
       conn.setRequestProperty(entry.getKey(), entry.getValue());