+++ firebase-firestore/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Unreleased
 * [feature] Expose Sum/Average aggregate query support in API. [#5217](//github.com/firebase/firebase-android-sdk/pull/5217)
+* [changed] Added Kotlin extensions (KTX) APIs from `com.google.firebase:firebase-firestore-ktx`
+  to `com.google.firebase:firebase-firestore` under the `com.google.firebase.firestore` package.
+  For details, see the
+  [FAQ about this initiative](https://firebase.google.com/docs/android/kotlin-migration)
+* [deprecated] All the APIs from `com.google.firebase:firebase-firestore-ktx` have been added to
+  `com.google.firebase:firebase-firestore` under the `com.google.firebase.firestore` package,
+  and all the Kotlin extensions (KTX) APIs in `com.google.firebase:firebase-firestore-ktx` are
+  now deprecated. As early as April 2024, we'll no longer release KTX modules. For details, see the
+  [FAQ about this initiative](https://firebase.google.com/docs/android/kotlin-migration)
+
 
 # 24.8.1
 * [fixed] Disabled `GradleMetadataPublishing` to fix breakage of the Kotlin extensions library. [#5337]